requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
motor==3.4.0
PyJWT==2.8.0
python-decouple==3.8